// src/posts/dto/post-author.dto.ts
export class PostAuthorDto {
    title: string;    
    price: number;    
    paymentType: string;
    description: string;      
    serviceType: string;    
    regions: string;    
    comuna: string;    
    image: string;      
    createdAt: Date;
    authorName: string;
    authorPhone: string;
    authorEmail: string;
}

    